ENH: Restructured model file structure.
