+++ hddm/models/__init__.py
@@ -1,0 +1,15 @@
+from base import AccumulatorModel, HDDMBase
+from hddm_truncated import HDDMTruncated
+from hddm_transformed import HDDM
+from hddm_stimcoding import HDDMStimCoding
+from hddm_regression import HDDMRegressor
+from hlba_truncated import HLBA
+
+__all__ = ['AccumulatorModel',
+           'HDDMBase',
+           'HDDMTruncated',
+           'HDDM',
+           'HDDMStimCoding',
+           'HDDMRegressor',
+           'HLBA'
+]
