+++ hddm/models/base.py
@@ -17,11 +17,9 @@ import pymc as pm
 
 import hddm
 import kabuki
-import kabuki.step_methods as steps
 import inspect
 
 from kabuki.hierarchical import Knode
-from copy import copy
 from scipy.optimize import fmin_powell, fmin
 
 class AccumulatorModel(kabuki.Hierarchical):
@@ -421,169 +419,6 @@ class HDDMBase(AccumulatorModel):
         #create the avg model
         avg_model  = self.__class__(self.data, include=self.include, is_group_model=False, **self._kwargs)
         return avg_model
-
-
-class HDDMTruncated(HDDMBase):
-    def _create_stochastic_knodes(self, include):
-        knodes = OrderedDict()
-        if 'a' in include:
-            knodes.update(self.create_family_trunc_normal('a', lower=1e-3, upper=1e3, value=1))
-        if 'v' in include:
-            knodes.update(self.create_family_normal('v', value=0))
-        if 't' in include:
-            knodes.update(self.create_family_trunc_normal('t', lower=1e-3, upper=1e3, value=.01))
-        if 'sv' in include:
-            # TW: Use kabuki.utils.HalfCauchy, S=10, value=1 instead?
-            knodes.update(self.create_family_trunc_normal('sv', lower=0, upper=1e3, value=1))
-            #knodes.update(self.create_family_exp('sv', value=1))
-        if 'sz' in include:
-            knodes.update(self.create_family_trunc_normal('sz', lower=0, upper=1, value=.1))
-        if 'st' in include:
-            knodes.update(self.create_family_trunc_normal('st', lower=0, upper=1e3, value=.01))
-        if 'z' in include:
-            knodes.update(self.create_family_trunc_normal('z', lower=0, upper=1, value=.5))
-        if 'p_outlier' in include:
-            knodes.update(self.create_family_trunc_normal('p_outlier', lower=0, upper=1, value=0.05))
-
-        return knodes
-
-class HDDM(HDDMBase):
-    def __init__(self, *args, **kwargs):
-        self.use_gibbs = kwargs.pop('use_gibbs_for_mean', True)
-        self.use_slice = kwargs.pop('use_slice_for_std', True)
-
-        super(HDDM, self).__init__(*args, **kwargs)
-
-    def pre_sample(self):
-        if not self.is_group_model:
-            return
-
-        # apply gibbs sampler to normal group nodes
-        for name, node_descr in self.iter_group_nodes():
-            node = node_descr['node']
-            knode_name = node_descr['knode_name'].replace('_trans', '')
-            if self.use_gibbs and isinstance(node, pm.Normal) and knode_name not in self.group_only_nodes:
-                self.mc.use_step_method(steps.kNormalNormal, node)
-            if self.use_slice and isinstance(node, pm.Uniform) and knode_name not in self.group_only_nodes:
-                self.mc.use_step_method(steps.UniformPriorNormalstd, node)
-
-    def _create_stochastic_knodes(self, include):
-        knodes = OrderedDict()
-        if 'a' in include:
-            knodes.update(self.create_family_exp('a', value=1))
-        if 'v' in include:
-            knodes.update(self.create_family_normal('v', value=0))
-        if 't' in include:
-            knodes.update(self.create_family_exp('t', value=.01))
-        if 'sv' in include:
-            # TW: Use kabuki.utils.HalfCauchy, S=10, value=1 instead?
-            knodes.update(self.create_family_trunc_normal('sv', lower=0, upper=1e3, value=1))
-            #knodes.update(self.create_family_exp('sv', value=1))
-        if 'sz' in include:
-            knodes.update(self.create_family_invlogit('sz', value=.1))
-        if 'st' in include:
-            knodes.update(self.create_family_exp('st', value=.01))
-        if 'z' in self.include:
-            knodes.update(self.create_family_invlogit('z', value=.5))
-        if 'p_outlier' in include:
-            knodes.update(self.create_family_invlogit('p_outlier', value=0.05))
-
-        return knodes
-
-    def _create_an_average_model(self):
-        """
-        create an average model for group model quantiles optimization.
-        """
-
-        #this code only check that the arguments are as expected, i.e. the constructor was not change
-        #since we wrote this function
-        super_init_function = super(self.__class__, self).__init__
-        init_args = set(inspect.getargspec(super_init_function).args)
-        known_args = set(['wiener_params', 'include', 'self', 'bias', 'data'])
-        assert known_args == init_args, "Arguments of the constructor are not as expected"
-
-        #create the avg model
-        avg_model  = self.__class__(self.data, include=self.include, is_group_model=False, **self._kwargs)
-        return avg_model
-
-
-class HDDMStimCoding(HDDM):
-    """HDDM model that can be used when stimulus coding and estimation
-    of bias (i.e. displacement of starting point z) is required.
-
-    In that case, the 'resp' column in your data should contain 0 and
-    1 for the chosen stimulus (or direction), not whether the response
-    was correct or not as you would use in accuracy coding. You then
-    have to provide another column (referred to as stim_col) which
-    contains information about which the correct response was.
-
-    :Arguments:
-        split_param : str ('v' or 'z')
-            There are two ways to model stimulus coding in the case where both stimuli
-            have equal information (so that there can be no difference in drift):
-            * 'z': Use z for stimulus A and 1-z for stimulus B
-            * 'v': Use drift v for stimulus A and -v for stimulus B
-
-        stim_col : str
-            Column name for extracting the stimuli to use for splitting.
-
-    """
-    def __init__(self, *args, **kwargs):
-        self.stim_col = kwargs.pop('stim_col', 'stim')
-        self.split_param = kwargs.pop('split_param', 'z')
-        if self.split_param == 'z' and 'include' in kwargs:
-            if 'z' not in kwargs['include']:
-                kwargs['include'].append('z')
-                print "Adding z to includes."
-        else:
-            kwargs['include'] = ['z']
-            print "Adding z to includes."
-        #assert self.stim_col in self.data.columns, "Can not find column named %s" % self.stim_col
-        self.stims = np.unique(args[0][self.stim_col])
-        assert len(self.stims) == 2, "%s must contain two stimulus types" % self.stim_col
-
-        super(HDDMStimCoding, self).__init__(*args, **kwargs)
-
-
-    def _create_wfpt_knode(self, knodes):
-        wfpt_parents = self._create_wfpt_parents_dict(knodes)
-        # Here we use a special Knode (see below) that either inverts v or z
-        # depending on what the correct stimulus was for that trial type.
-        return KnodeWfptStimCoding(self.wfpt_class, 'wfpt',
-                                   observed=True, col_name='rt',
-                                   depends=[self.stim_col],
-                                   split_param=self.split_param,
-                                   stims=self.stims,
-                                   stim_col=self.stim_col,
-                                   **wfpt_parents)
-
-class KnodeWfptStimCoding(Knode):
-    def __init__(self, *args, **kwargs):
-        self.split_param = kwargs.pop('split_param')
-        self.stims = kwargs.pop('stims')
-        self.stim_col = kwargs.pop('stim_col')
-        super(KnodeWfptStimCoding, self).__init__(*args, **kwargs)
-
-    def create_node(self, name, kwargs, data):
-        # the addition of "depends=['stim']" in the call of
-        # KnodeWfptInvZ in HDDMStimCoding makes that data are
-        # submitted splitted by the values of the variable stim the
-        # following lines check if the variable stim is equal to the
-        # value of stim for which z' = 1-z and transforms z if this is
-        # the case (similar to v)
-        if all(data[self.stim_col] == self.stims[0]):
-            if self.split_param == 'z':
-                z = copy(kwargs['z'])
-                kwargs['z'] = 1-z
-            elif self.split_param == 'v':
-                v = copy(kwargs['v'])
-                kwargs['v'] = -v
-            else:
-                raise ValueError('split_var must be either v or z, but is %s' % self.split_var)
-
-            return self.pymc_node(name, **kwargs)
-        else:
-            return self.pymc_node(name, **kwargs)
 
 if __name__ == "__main__":
     import doctest
